CheerEvent Fix UI Improvement

+++ src/main/java/de/MCmoderSD/objects/Timer.java
@@ -69,7 +69,7 @@ public class Timer {
                 channel,
                 channel,
                 response,
-                null, null, null), "Timer: " + name, response);
+                null, (Integer) null, null), "Timer: " + name, response);
     }
 
     // Trigger
